+++ src/main/java/org/realityforge/replicant/server/ee/EntityMessageCacheUtil.java
@@ -3,6 +3,8 @@ package org.realityforge.replicant.server.ee;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import javax.naming.InitialContext;
+import javax.naming.NamingException;
 import javax.transaction.TransactionSynchronizationRegistry;
 import org.realityforge.replicant.server.EntityMessageSet;
 
@@ -20,8 +22,20 @@ public final class EntityMessageCacheUtil
    */
   private static final String SESSION_KEY = KEY + "/Session";
 
+  /**
+   * Standard JNDI key for TransactionSynchronizationRegistry.
+   */
+  private static final String REGISTRY_KEY = "java:comp/TransactionSynchronizationRegistry";
+
+
   private EntityMessageCacheUtil()
   {
+  }
+
+  @Nonnull
+  public static EntityMessageSet getEntityMessageSet()
+  {
+    return getEntityMessageSet( lookupTransactionSynchronizationRegistry(), KEY );
   }
 
   @Nonnull
@@ -31,15 +45,33 @@ public final class EntityMessageCacheUtil
   }
 
   @Nullable
+  public static EntityMessageSet lookupEntityMessageSet()
+  {
+    return lookupEntityMessageSet( lookupTransactionSynchronizationRegistry(), KEY );
+  }
+
+  @Nullable
   public static EntityMessageSet lookupEntityMessageSet( @Nonnull final TransactionSynchronizationRegistry r )
   {
     return lookupEntityMessageSet( r, KEY );
   }
 
   @Nullable
+  public static EntityMessageSet removeEntityMessageSet()
+  {
+    return removeEntityMessageSet( lookupTransactionSynchronizationRegistry(), KEY );
+  }
+
+  @Nullable
   public static EntityMessageSet removeEntityMessageSet( @Nonnull final TransactionSynchronizationRegistry r )
   {
     return removeEntityMessageSet( r, KEY );
+  }
+
+  @Nonnull
+  public static EntityMessageSet getSessionEntityMessageSet()
+  {
+    return getEntityMessageSet( lookupTransactionSynchronizationRegistry(), SESSION_KEY );
   }
 
   @Nonnull
@@ -55,6 +87,18 @@ public final class EntityMessageCacheUtil
   }
 
   @Nullable
+  public static EntityMessageSet lookupSessionEntityMessageSet()
+  {
+    return lookupEntityMessageSet( lookupTransactionSynchronizationRegistry(), SESSION_KEY );
+  }
+
+  @Nullable
+  public static EntityMessageSet removeSessionEntityMessageSet()
+  {
+    return removeEntityMessageSet( lookupTransactionSynchronizationRegistry(), SESSION_KEY );
+  }
+
+  @Nullable
   public static EntityMessageSet removeSessionEntityMessageSet( @Nonnull final TransactionSynchronizationRegistry r )
   {
     return removeEntityMessageSet( r, SESSION_KEY );
@@ -63,7 +107,7 @@ public final class EntityMessageCacheUtil
   private static EntityMessageSet removeEntityMessageSet( final TransactionSynchronizationRegistry r, final String key )
   {
     final EntityMessageSet messageSet = lookupEntityMessageSet( r, key );
-    if( null != messageSet )
+    if ( null != messageSet )
     {
       r.putResource( key, null );
     }
@@ -73,7 +117,7 @@ public final class EntityMessageCacheUtil
   private static EntityMessageSet getEntityMessageSet( final TransactionSynchronizationRegistry r, final String key )
   {
     EntityMessageSet messageSet = lookupEntityMessageSet( r, key );
-    if( null == messageSet )
+    if ( null == messageSet )
     {
       messageSet = new EntityMessageSet();
       r.putResource( key, messageSet );
@@ -84,5 +128,20 @@ public final class EntityMessageCacheUtil
   private static EntityMessageSet lookupEntityMessageSet( final TransactionSynchronizationRegistry r, final String key )
   {
     return (EntityMessageSet) r.getResource( key );
+  }
+
+  @Nonnull
+  private static TransactionSynchronizationRegistry lookupTransactionSynchronizationRegistry()
+  {
+    try
+    {
+      return (TransactionSynchronizationRegistry) new InitialContext().lookup( REGISTRY_KEY );
+    }
+    catch ( final NamingException ne )
+    {
+      final String message =
+        "Unable to locate TransactionSynchronizationRegistry at " + REGISTRY_KEY + " due to " + ne;
+      throw new IllegalStateException( message, ne );
+    }
   }
 }