+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.5.7:
+* Enhance EntityMessageCacheUtil to support variants of method that lookup
+  TransactionSynchronizationRegistry in JNDI.
 * Enhance EntityMessageSet to support merging collections of EntityMessage.
 * Enhance EntityMessageSet to support optional cloning of EntityMessage instances
   on initial merge.